LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-snap/snap-after-style-change-respects-scroll-behavior-expected.txt



FAIL Dynamically enabling scroll-snap should trigger a smooth scroll if scroll-behavior is smooth. assert_equals: Scroller should be at the initial test position. expected 160 but got 1
